example/requirements.txt
django-teryt
